unibus read/write clean-up

+++ bus.h
@@ -156,7 +156,7 @@ public:
 	uint16_t readWord(const uint16_t a, const d_i_space_t s = i_space);
 	uint16_t peekWord(const uint16_t a);
 
-	uint16_t readUnibusByte(const uint32_t a);
+	uint8_t  readUnibusByte(const uint32_t a);
 
 	write_rc_t write    (const uint16_t a, const word_mode_t word_mode, uint16_t value, const rm_selection_t mode_selection, const d_i_space_t s = i_space);
 	void       writeByte(const uint16_t a, const uint8_t value) { write(a, wm_byte, value, rm_cur); }
